// src/core/types/DFBlockCodename.ts
/**
 * Raw DiamondFire codeblock codename.
 */
type DFBlockCodename =
	| "bracket"
	| "event"
	| "player_action"
	| "if_player"
	| "entity_event"
	| "entity_action"
	| "if_entity"
	| "set_var"
	| "if_var"
	| "select_obj"
	| "game_action"
	| "if_game"
	| "func"
	| "call_func"
	| "process"
	| "start_process"
	| "control"
	;

export default DFBlockCodename;